Jank now compiles as a static lib - No warnings yay!

+++ Jank/src/Jank/ImGui/ImGuiLayer.cpp
@@ -66,12 +66,11 @@ namespace Jank {
 		ImGui_ImplGlfw_NewFrame();
 		ImGui::NewFrame();
 	}
-
 	void ImGuiLayer::End()
 	{
 		ImGuiIO& io = ImGui::GetIO();
 		Application& app = Application::Get();
-		io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+		io.DisplaySize = ImVec2((float)app.GetWindow().GetWidth(), (float)app.GetWindow().GetHeight());
 
 		//Rendering
 		ImGui::Render();
