+++ Jank/src/Platform/Windows/WindowsWindow.cpp
@@ -144,7 +144,7 @@ namespace Jank
 		{
 			WindowData& data = *(WindowData*)glfwGetWindowUserPointer(window);
 
-			MouseMovedEvent event(xPos, yPos);
+			MouseMovedEvent event((float)xPos, (float)yPos);
 			data.EventCallback(event);
 		});
 	}
